requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
pydantic>=2.10.0
google-cloud-firestore>=2.19.0
firebase-admin>=6.6.0
